+++ common/javascript/main.js
@@ -2,10 +2,12 @@ import govukFrontend from 'govuk-frontend/govuk/all';
 
 import missedMilestonesChart from './missed-milestones';
 import upcomingMilestoneChart from './upcoming-milestones';
+import removeAccordionCross from './remove-accordion-cross';
 
 window.TRANSITIONDELIVERYDASHBOARD = {
   missedMilestonesChart,
-  upcomingMilestoneChart
+  upcomingMilestoneChart,
+  removeAccordionCross
 };
 
 document.addEventListener('DOMContentLoaded', function() {
@@ -16,22 +18,6 @@ document.addEventListener('DOMContentLoaded', function() {
 // These will be removed from the session storage so the accordions will collapse rather than stay open.
 sessionStorage.clear();
 
-// Removes the plus icon from the accordion if the table is empty 
-const accordionTable = document.getElementById('accordion-table');
-const milestoneTable = document.getElementsByClassName('table-content');
-
-let count = 0;
-
-for (var i = 0; i < milestoneTable.length; i++) {
-  if (milestoneTable[i].rows.length == 0) {
-    const accordionHeader = milestoneTable[i].parentElement.parentElement.querySelector('.govuk-accordion__section-heading');
-    accordionHeader.classList.add('remove-icon')
-    count++;
-  }
-}
-if (count == milestoneTable.length) {
-  accordionTable.classList.add('remove-open-all')
-}
